+++ BE/src/main/java/louie/hanse/issuetracker/exception/MemberNotFoundException.java
@@ -1,0 +1,8 @@
+package louie.hanse.issuetracker.exception;
+
+public class MemberNotFoundException extends RuntimeException {
+
+    public MemberNotFoundException(String message) {
+        super(message);
+    }
+}
